Added a simple FOV
Basically a Field Of View Mesh that follows the player and the cursor, making the DarkWood "Fog Of War" Illusion. I tried adding a colliding function but it messes up the positioning of the FOV, and considering that there are yet to be any colliders in the game, i decided to stick with the code without the collider part (It is still included in the script "Field Of View" but as a comment).